Restructured the project to clean up the project root Fixed a missing `extern` which was causing compilation to fail Replaced CMake with Make Moved new build info to README Added gitignore to keep unwanted files out of the tree
