--- INSTALL
@@ -1,31 +1,0 @@
-
-unix:
-
-
-to compile ppm type the following in the ppm-directory:
-
-make
-
-to install the binaries into /usr/local/bin/ type:
-
-make install
-
-
-windows:
-
-to compile ppm add "-DWIN32" (defines WIN32) to the makefile and type
-
-make
-
-or if you have visual studio then you can open the project/
-workspace files for ppm that come with the archive and use vs to
-do the dirty work for you.
-
-
-if compiling/linking fails, try to give (if you are using bash)
- 
-export LD=gcc
-export CC=gcc
- 
-and try again. otherwise you'll have to edit the makefile.
-
