+++ src/color.c
@@ -11,12 +11,12 @@
 #include <gtk/gtkwidget.h>
 #include <gdk/gdkkeysyms.h>
 
-#include "defines.h"
-#include "main.h"
-#include "draw.h"
-#include "color.h"
-#include "refresh.h"
-#include "mouse.h"
+#include "include/defines.h"
+#include "include/main.h"
+#include "include/draw.h"
+#include "include/color.h"
+#include "include/refresh.h"
+#include "include/mouse.h"
 
 
 extern unsigned char memory_data[8 * 8 * 1024]; /* one byte / pixel */
