+++ src/include/edit.h
