+++ src/memory.c
@@ -11,11 +11,11 @@
 #include <gtk/gtkwidget.h>
 #include <gdk/gdkkeysyms.h>
 
-#include "defines.h"
-#include "main.h"
-#include "draw.h"
-#include "memory.h"
-#include "button.h"
+#include "include/defines.h"
+#include "include/main.h"
+#include "include/draw.h"
+#include "include/memory.h"
+#include "include/button.h"
 
 
 extern unsigned char memory_data[8 * 8 * 1024]; /* one byte / pixel */
