+++ src/png.c
@@ -7,7 +7,7 @@
 
 #include <png.h>
 
-#include "defines.h"
+#include "include/defines.h"
 #include "png.h"
